# Use garbage-free formatter for `s` and `S` patterns