--- a/log4j-core-test/src/test/java/org/apache/logging/log4j/core/util/internal/instant/InstantPatternDynamicFormatterTest.java
+++ b/log4j-core-test/src/test/java/org/apache/logging/log4j/core/util/internal/instant/InstantPatternDynamicFormatterTest.java
@@ -322,9 +322,9 @@ class InstantPatternDynamicFormatterTest {
         final MutableInstant instant = new MutableInstant();
         // In the 1970's some time zones had sub-minute offsets to UTC, e.g., Africa/Monrovia.
         // We will exclude them for tests:
-        final long minEpochSecond = 315_532_800; // 1980-01-01 01:00:00
-        final long maxEpochSecond = 1_621_280_470; // 2021-05-17 21:41:10
-        final long epochSecond = minEpochSecond + RANDOM.nextLong(maxEpochSecond - minEpochSecond);
+        final int minEpochSecond = 315_532_800; // 1980-01-01 01:00:00
+        final int maxEpochSecond = 1_621_280_470; // 2021-05-17 21:41:10
+        final long epochSecond = minEpochSecond + RANDOM.nextInt(maxEpochSecond - minEpochSecond);
         final int epochSecondNano = randomNanos();
         instant.initFromEpochSecond(epochSecond, epochSecondNano);
         return instant;