--- a/log4j-core/src/main/java/org/apache/logging/log4j/core/util/internal/instant/InstantPatternDynamicFormatter.java
+++ b/log4j-core/src/main/java/org/apache/logging/log4j/core/util/internal/instant/InstantPatternDynamicFormatter.java
@@ -351,7 +351,7 @@ final class InstantPatternDynamicFormatter implements InstantPatternFormatter {
                 : java.time.Instant.ofEpochSecond(instant.getEpochSecond(), instant.getNanoOfSecond());
     }
 
-    abstract static class AbstractFormatter implements InstantPatternFormatter {
+    private abstract static class AbstractFormatter implements InstantPatternFormatter {
 
         private final String pattern;